에피소드 객체 생성완료
